+++ src/components/Setting/style.tsx
@@ -1,24 +1,15 @@
 import styled from 'styled-components';
+
 import theme from '@/styles/theme';
 
-export const Container = styled.div`
-  width: 100vw;
-  height: 100vh;
-  background-color: ${theme.colors.primary[100]};
-`;
-
-export const CardWrapper = styled.div`
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  width: 80vw;
-  max-width: 600px;
-  padding: 2rem 2rem 4rem 4rem;
-  background-color: white;
+export const ContentWrapper = styled.div`
   display: flex;
   flex-direction: column;
+  align-items: flex-end;
   gap: 2rem;
+  background-color: ${theme.isDark
+    ? theme.colors.grayscale[400]
+    : theme.colors.white};
 `;
 
 export const ButtonWrapper = styled.div`
@@ -26,25 +17,18 @@ export const ButtonWrapper = styled.div`
   justify-content: flex-end;
 `;
 
-export const ContentContainer = styled.div`
-  display: flex;
-  gap: 1rem;
-`;
-
-export const AvatarWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-`;
-
 export const InputWrapper = styled.div`
   display: flex;
   flex-direction: column;
-  gap: 1rem;
+  gap: 2rem;
 `;
 
 export const TextWrapper = styled.div`
+  position: absolute;
+  bottom: 30px;
+  right: 15px;
   display: flex;
-  align-items: flex-end;
+  align-items: center;
+  justify-content: center;
   gap: 1rem;
 `;
