+++ src/components/common/Dropdown/style.tsx
@@ -65,7 +65,7 @@ export const ListItemButton = styled.button`
 
 export const DropdownContent = styled.ul`
   width: 100%;
-  min-width: 100px;
+  min-width: 120px;
   position: absolute;
   top: 100%;
   margin: 0px;
